D8dashboard/GotionBattery/requirements.txt
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
et-xmlfile==1.0.1
idna==2.8
influxdb==5.2.3
jdcal==1.4.1
kiwisolver==1.1.0
lxml==4.4.1
matplotlib==3.1.1
numpy==1.17.2
Office365-REST-Python-Client==2.1.3
openpyxl==2.6.3
pandas==0.25.1
Pillow==6.1.0
PyMySQL==0.9.3
pyparsing==2.4.2
python-dateutil==2.8.0
python-docx==0.8.10
pytz==2019.2
requests==2.22.0
six==1.12.0
urllib3==1.25.6
wxPython==4.0.6
